Modified BlackjackPlayer to subtract bets from available cash, and add appropriate amounts back on win.
View no longer performs calculation.

Realized that for some insane reason all bets were represented as integers--along with available cash--so changed Bet and Cash to decimal type.

Updated associated methods and calculations to work with decimal type instead of int.

+++ Blackjack/Application/BlackjackController.cs
@@ -23,7 +23,7 @@ namespace Blackjack.Application
             switch (args[0])
             {
                 case "bet":
-                    if (int.TryParse(args[1], out int bet))
+                    if (decimal.TryParse(args[1], out decimal bet))
                         model.Bet(bet);
                     break;
                 case "hit":
@@ -32,7 +32,7 @@ namespace Blackjack.Application
                 case "stand":
                     model.Stand();
                     break;
-                case "double":
+                case "decimal":
                     model.DoubleDown();
                     break;
                 case "split":
